requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
requests==2.31.0
joblib==1.3.2
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.0.3
pillow==10.0.0
python-multipart==0.0.6